requirements.txt
pymongo==4.10.1
pandas==2.2.0
jupyter==1.0.0
ipykernel==6.29.4
seaborn==0.13.1
python-dotenv
scipy
fastapi
uvicorn
neo4j
